Return firmware version parts as uint8_t instead of int
The firmware version parts returned by the Nitrokey devices are uint8_t
values.  This is part of the public API as part of the NK_storage_status
struct.  For consistency with this API, this patch changes the functions
NK_get_major_firmware_version and NK_get_minor_firmware_version to
return uint8_t instead of int.

+++ NK_C_API.cc
@@ -719,14 +719,14 @@ NK_C_API char* NK_get_SD_usage_data_as_string() {
 		}, -2));
 	}
 
-	NK_C_API int NK_get_major_firmware_version() {
+	NK_C_API uint8_t NK_get_major_firmware_version() {
 		auto m = NitrokeyManager::instance();
 		return get_with_result([&]() {
 			return m->get_major_firmware_version();
 		});
 	}
 
-  NK_C_API int NK_get_minor_firmware_version() {
+  NK_C_API uint8_t NK_get_minor_firmware_version() {
 		auto m = NitrokeyManager::instance();
 		return get_with_result([&]() {
 			return m->get_minor_firmware_version();
